pyproject.toml
[project]
name = "chatbot-langgraph"
version = "0.1.0"
description = "Multi-step chatbot using LangGraph"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
